feat: added global move entities message allowed 1000+ entities perfectly synchronized over 4 client

+++ Server/include/LKZ/Core/ECS/Entity.h
@@ -4,6 +4,6 @@
 /**
  * @brief Type representing an entity ID.
  */
-using Entity = uint32_t;
+using Entity = uint16_t;
 
 constexpr Entity INVALID_ENTITY = 0;
