+++ Server/src/Protocol/Message/Entity/MoveEntityMessage.cpp
@@ -1,6 +1,6 @@
 #include "LKZ/Protocol/Message/Entity/MoveEntityMessage.h"
 
-MoveEntityMessage::MoveEntityMessage(int entityId, float x, float y, float z)
+MoveEntityMessage::MoveEntityMessage(uint16_t entityId, float x, float y, float z)
 {
 
     this->entityId = entityId;
@@ -10,7 +10,7 @@ MoveEntityMessage::MoveEntityMessage(int entityId, float x, float y, float z)
 }
 MoveEntityMessage::MoveEntityMessage() {};
 
-byte MoveEntityMessage::getId() const
+uint8_t MoveEntityMessage::getId() const
 {
     return ID;
 }
@@ -18,7 +18,7 @@ byte MoveEntityMessage::getId() const
 std::vector<uint8_t>& MoveEntityMessage::serialize(Serializer& serializer) const
 {
     serializer.writeByte(ID);
-    serializer.writeInt(entityId);
+    serializer.writeUInt16(entityId);
     serializer.writeFloat(posX);
     serializer.writeFloat(posY);
     serializer.writeFloat(posZ);
